move picklist.py from sourmash.sig into sourmash
